# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: you@your-company.com
  description: This is a simple API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AL
  version: 1.0.1
servers:
- description: SwaggerHub API Auto Mocking
  url: https://accelerated-life-demo.herokuapp.com/api
...
paths:
...
  /auctions/auction:
    delete:
      description: delete an auction. If it is of type remove then repay the old winner.
        If it is of type resolve or other than remove, then just remove the auction.
      operationId: deleteAuction
      parameters:
      - description: pass the _id to be deleted
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Auction'
                type: array
          description: auctions matching criteria
        "404":
          description: Auction not found
      summary: delete auction
      tags:
      - Staff
    get:
      description: by passing in a auction ID that auction is returned
      operationId: searchAuctions
      parameters:
      - description: pass the id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Auction'
                type: array
          description: auction matching criteria
        "404":
          description: No auction found
      summary: searches auctions
      tags:
      - AllSchool
  /auctions:
    get:
      description: a teacher asks for all their auctions. Winner, startdate, enddate,
        itemNumber, bid, description, visibility is returned. Winner is populated
        with firstName and lastName and sorted by endDate ascending. Visibility is
        populated with name and sorted by name ascending.
      operationId: searchAuctionsTeacher
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Auction'
                type: array
          description: auctions matching criteria
        "404":
          description: No auctions found
      summary: searches auctions
      tags:
      - Staff
    post:
      description: when a new auction is posted check to see if it visibility belongs
        to the major groups with its period being 0,9,10,11,12. If it does add those
        groups to visibility and save the auction. If it does not check visibility
        == 1 or all the classes of the teacher. If it does then find all the classes
        of the teacher and add it to the visibility and save. Finally, it is not of
        any major group and that means that the teacher is auctioning to some of their
        classes, add those class id's to the visibility and auction. For all situations
        return all the auctions ownened by the user. With visibility populated with
        name and sorted by name ascending. With winner populated with firstName and
        lastName. Make all auctions sorted by endDate ascending.
      operationId: makeAuction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestMakeAuction'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Auction'
                type: array
          description: auctions matching criteria
        "400":
          description: Auction cannot be made with that data
      summary: make a new auction
      tags:
      - Staff
...
  /classes/class:
    delete:
      description: delete a class
      operationId: deleteclass
      parameters:
      - description: pass the _id to be deleted
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          description: School Deleted
        "404":
          description: Class cannot be deleted with that data
      summary: delete class
      tags:
      - Staff
    get:
      description: By passing in a _id you can find a class. Members should be populated
        with firstName, lastName, rank, netWorth
      operationId: searchClass
      parameters:
      - description: pass the _id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassWithMembers'
          description: class matching criteria
        "404":
          description: No class found
      summary: searches for a class
      tags:
      - All
    post:
      description: creates a new class
      operationId: makeClass
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestMakeClass'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Class'
                type: array
          description: class record is made. Return all the classes from this owner.
            Sort classes by period ascending
        "404":
          description: Class cannot be made with that data
      summary: make a new class
      tags:
      - Staff
    put:
      description: edit a class. return all the classes of the owner. Exclude owner,
        createdAt, updatedAt and version. Sort by period ascending
      operationId: editClass
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestEditClass'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
          description: school record is edited
        "404":
          description: Class cannot be edited with that data
      summary: edit class
      tags:
      - Staff
  /classes/class/kick:
    put:
      description: The class owner is kicking a class member. Remove the kicked members
        _id from the class members list
      operationId: kickClass
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestKickClass'
      responses:
        "200":
          description: class member has been removed from class
        "404":
          description: member cannot be removed with that data
      summary: Kick member from class
      tags:
      - Staff
...
  /classes:
    get:
      description: By passing in a user_id you can find classes. Sort by period ascending.
      operationId: searchClasses
      parameters:
      - description: pass the user_id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Class'
                type: array
          description: class matching criteria
        "404":
          description: No classes found
      summary: searches for users classes
      tags:
      - Staff
...
  /events:
    get:
      description: first get all the users that share the same school as the staff.
        then get all the events that are today or greater and the owner_id is from
        the list of users found previously
      operationId: searchEvents
      parameters:
      - description: pass the school _id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ResponseEvents'
                type: array
          description: events matching criteria in value order decending
        "404":
          description: No events matching that school _id
      summary: returns all of todays events for the users school
      tags:
      - Staff
  /transactions:
    get:
      description: By passing in the staffs _id you can find a transactions that they
        are owner of. Instead of owner_id put object.student firstName lastName from
        populated field
      operationId: searchTransactions
      parameters:
      - description: pass the staff_id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ResponseTransactions'
                type: array
          description: transactions matching criteria
        "404":
          description: No transaction found
      summary: all transactions a teacher is the owner of. Limit 30 sort by createdAt
        descending
      tags:
      - Staff
...
  /transactions/payTransaction:
    post:
      description: Make a new transaction. Find the associated buck and account. If
        there is no account then make one. Use algorithm to manage totalCurrency and
        freeCurrency on the buck(aka central bank). Then manage the change on account
        value. Accounts could be negative or amount coming in could be negative.
      operationId: payTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPayTransaction'
      responses:
        "200":
          description: return success status
        "404":
          description: transaction cannot be made with that data
      summary: When a teacher or admin is paying/debting a student with their own
        bucks
      tags:
      - Staff
  /transactions/payTransactions:
    post:
      description: Make a new transaction. Find the associated buck and account. If
        there is no account then make one. Use algorithm to manage totalCurrency and
        freeCurrency on the buck(aka central bank). Then manage the change on account
        value. Accounts could be negative or amount coming in could be negative. Do
        the above for every student in a class.
      operationId: payTransactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPayTransactions'
      responses:
        "200":
          description: return success status
        "400":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ResponsePayTransactions'
                type: array
          description: at lest 1 transaction failed
      summary: When a teacher or admin is paying/debting an entire class
      tags:
      - Staff
...
  /users/user:
    delete:
      description: Delete a student.
      operationId: deleteStudent
      parameters:
      - description: pass the email to be deleted. Delete all of their transactions
          and events. Search classes that the user is a member in and remove them
          from the member list. Find all of their buck accounts and burn the bucks
          then delete the account. Find all the crypto accounts, modify the cryptos
          totalCount then delete the account
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          description: transaction Deleted
        "404":
          description: There is no user with that email
      summary: delete student
      tags:
      - Staff
    get:
      description: return student
      operationId: searchStudent
      parameters:
      - description: pass the _id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: students matching criteria
        "404":
          description: No user found with that data
      summary: return one student
      tags:
      - All
    put:
      description: edit user fields. If it is password then the password must be salted
        and hashed. If it is college or career then income needs to be halved and
        end times need to be saved to the record. Email change means verified is set
        to false. Request will have new password in the clear. If college true then
        transaction has to be made representing the cost of college.
      operationId: userEdit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/users_user_body'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: return user information with history
        "404":
          description: Cannot update user with that data
      summary: edit a user
      tags:
      - All
...
  /users/resetPassword:
    post:
      description: generate temp password and send it in an email
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUser'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseResetPassword'
          description: class record is made. Return all the classes from this owner.
            Sort classes by period ascending
        "401":
          description: unconfirmed email
      summary: reset password
      tags:
      - Staff
...
components:
  schemas:
...
    Auction:
      example:
        visibility:
        - 613eab19e58cc511801af657
        - 613eab19e58cc511801af657
        endDate: endDate
        owner_id:
          lastName: Jones
          _id: ou812
        maxBid: 4
        winner_id:
          firstName: Tom
          lastName: Jones
          _id: ou812
        description: Candy Bar
        _id: 613eab19e58cc511801af657
        bid: 4
        startDate: startDate
      properties:
        _id:
          example: 613eab19e58cc511801af657
          type: string
        owner_id:
          $ref: '#/components/schemas/Auction_owner_id'
        winner_id:
          $ref: '#/components/schemas/Auction_winner_id'
        startDate:
          type: string
        endDate:
          type: string
        bid:
          default: 0
          example: 4
          type: integer
        maxBid:
          default: 0
          example: 4
          type: integer
        description:
          example: Candy Bar
          type: string
        visibility:
          items:
            example: 613eab19e58cc511801af657
            type: string
          type: array
      required:
      - EndDate
      - _id
      - bid
      - description
      - owner
      - startDate
      - visibility
      - winner
      type: object
...
    Class:
      example:
        period: 4
        owner_id: 613eab19e58cc511801af657
        members:
        - 613eab19e58cc511801af657
        - 613eab19e58cc511801af657
        name: Math
        _id: 613eab19e58cc511801af657
        addCode: Sd45Tb
      properties:
        _id:
          example: 613eab19e58cc511801af657
          type: string
        owner_id:
          example: 613eab19e58cc511801af657
          type: string
        period:
          example: 4
          type: integer
        name:
          example: Math
          type: string
        addCode:
          example: Sd45Tb
          type: string
        members:
          items:
            example: 613eab19e58cc511801af657
            type: string
          type: array
      required:
      - _id
      - addCode
      - members
      - name
      - owner_id
      - period
      type: object
    ClassWithMembers:
      example:
        period: 4
        owner_id: 613eab19e58cc511801af657
        members:
        - firstName: Jay
          lastName: Leno
          netWorth: 22.1
          rank: 4
          _id: 613eab19e58cc511801af657
        - firstName: Jay
          lastName: Leno
          netWorth: 22.1
          rank: 4
          _id: 613eab19e58cc511801af657
        name: Math
        _id: 613eab19e58cc511801af657
        addCode: Sd45Tb
      properties:
        _id:
          example: 613eab19e58cc511801af657
          type: string
        owner_id:
          example: 613eab19e58cc511801af657
          type: string
        period:
          example: 4
          type: integer
        name:
          example: Math
          type: string
        addCode:
          example: Sd45Tb
          type: string
        members:
          items:
            $ref: '#/components/schemas/ClassWithMembers_members'
          type: array
      required:
      - _id
      - addCode
      - members
      - name
      - owner_id
      - period
      type: object
...
    ResponseEvents:
      example:
        transaction_id: 613eab19e58cc511801af657
        createdAt: 2000-01-23T04:56:07.000+00:00
        owner_id:
          firstName: Jay
          lastName: Leno
        description: on time
        value: 123
      properties:
        value:
          example: 123
          type: integer
        createdAt:
          format: date-time
          type: string
        description:
          example: on time
          type: string
        transaction_id:
          example: 613eab19e58cc511801af657
          type: string
        owner_id:
          $ref: '#/components/schemas/ResponseEvents_owner_id'
      type: object
...
    ResponsePayTransactions:
      properties:
        message:
          example: ss1@ss.com was not paid 2300
          type: string
      type: object
...
    ResponseTransactions:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        amount: 123.45
        student: Jay Leno
        description: on time
      properties:
        amount:
          example: 123.45
          format: float
          type: number
        createdAt:
          format: date-time
          type: string
        description:
          example: on time
          type: string
        student:
          example: Jay Leno
          type: string
      type: object
...
    RequestEditClass:
      example:
        period: 4
        name: Math
        _id: 613eab19e58cc511801af657
      properties:
        name:
          example: Math
          type: string
        period:
          example: 4
          type: integer
        _id:
          example: 613eab19e58cc511801af657
          type: string
      type: object
    RequestKickClass:
      example:
        kick_id: 613eab19e58cc511801af657
        _id: 613eab19e58cc511801af657
      properties:
        kick_id:
          example: 613eab19e58cc511801af657
          type: string
        _id:
          example: 613eab19e58cc511801af657
          type: string
      type: object
    RequestMakeAuction:
      example:
        visibility:
        - 613eab19e58cc511801af657
        - 613eab19e58cc511801af657
        endDate: 2000-01-23T04:56:07.000+00:00
        owner_id: 61be17e2f5691463b0fc972d
        maxBid: 123
        description: candy
        bid: 123
        startDate: 2000-01-23T04:56:07.000+00:00
      properties:
        bid:
          example: 123
          type: number
        maxBid:
          example: 123
          type: number
        description:
          example: candy
          type: string
        endDate:
          format: date-time
          type: string
        startDate:
          format: date-time
          type: string
        owner_id:
          example: 61be17e2f5691463b0fc972d
          type: string
        visibility:
          items:
            example: 613eab19e58cc511801af657
            type: string
          type: array
      type: object
    RequestMakeClass:
      example:
        period: 4
        owner_id: 613eab19e58cc511801af657
        name: Math
      properties:
        name:
          example: Math
          type: string
        owner_id:
          example: 613eab19e58cc511801af657
          type: string
        period:
          example: 4
          type: integer
      type: object
...
    RequestPayTransaction:
      example:
        amount: 123.45
        student: 614290257145de7a9e40c58d
        owner_id: 613eab19e58cc511801af657
        description: Life Event
      properties:
        owner_id:
          example: 613eab19e58cc511801af657
          type: string
        description:
          example: Life Event
          type: string
        amount:
          example: 123.45
          format: float
          type: number
        student:
          example: 614290257145de7a9e40c58d
          type: string
      type: object
    RequestPayTransactions:
      example:
        owner: 613eab19e58cc511801af657
        amount: 123.45
        description: Life Event
        students:
        - 614290257145de7a9e40c58d
        - 614290257145de7a9e40c58d
      properties:
        owner:
          example: 613eab19e58cc511801af657
          type: string
        description:
          example: Life Event
          type: string
        amount:
          example: 123.45
          format: float
          type: number
        students:
          items:
            example: 614290257145de7a9e40c58d
            type: string
          type: array
      type: object
...
    ResponseResetPassword:
      example:
        password: wetSwd
      properties:
        password:
          example: wetSwd
          type: string
      required:
      - password
      type: object
    RequestUser:
      example:
        _id: 613eab19e58cc511801af657
      properties:
        _id:
          example: 613eab19e58cc511801af657
          type: string
      required:
      - _id
      type: object
...
    User:
      allOf:
      - $ref: '#/components/schemas/UserNoHistory'
      example:
        history:
        - date: 2000-01-23T04:56:07.000+00:00
          netWorth: 123.45
        - date: 2000-01-23T04:56:07.000+00:00
          netWorth: 123.45
      properties:
        history:
          items:
            $ref: '#/components/schemas/History'
          type: array
      required:
      - history
      type: object
...
    users_user_body:
      properties:
        email:
          example: ss@ss.com
          format: email
          type: string
        firstName:
          example: Tom
          type: string
        lastName:
          example: Jones
          type: string
        password:
          example: passw0rd
          type: string
        college:
          example: true
          type: boolean
        careerTransition:
          example: true
          type: boolean
      type: object
...
    Auction_owner_id:
      example:
        lastName: Jones
        _id: ou812
      properties:
        lastName:
          example: Jones
          type: string
        _id:
          example: ou812
          type: string
      type: object
    Auction_winner_id:
      example:
        firstName: Tom
        lastName: Jones
        _id: ou812
      properties:
        firstName:
          example: Tom
          type: string
        lastName:
          example: Jones
          type: string
        _id:
          example: ou812
          type: string
      type: object
...